Moved types to a separate file
